+++ android/src/main/java/com/rnmaps/maps/MapAirModulePackage.java
@@ -9,11 +9,14 @@ import com.facebook.react.module.model.ReactModuleInfo;
 import com.facebook.react.module.model.ReactModuleInfoProvider;
 import com.facebook.react.uimanager.ViewManager;
 import com.rnmaps.fabric.CalloutManager;
+import com.rnmaps.fabric.CircleManager;
 import com.rnmaps.fabric.MapViewManager;
 import com.rnmaps.fabric.MarkerManager;
 import com.rnmaps.fabric.NativeAirMapsModule;
 import com.rnmaps.fabric.NativeAirMapsModuleSpec;
+import com.rnmaps.fabric.OverlayManager;
 import com.rnmaps.fabric.PolygonManager;
+import com.rnmaps.fabric.PolylineManager;
 
 import java.util.HashMap;
 import java.util.List;
@@ -23,12 +26,21 @@ public class MapAirModulePackage extends TurboReactPackage {
 
     @Override
     public List<ViewManager> createViewManagers(ReactApplicationContext reactContext) {
-        return List.of(new MapViewManager(reactContext), new MarkerManager(reactContext), new CalloutManager(reactContext), new PolygonManager(reactContext));
+        return List.of(new MapViewManager(reactContext), new MarkerManager(reactContext), new CalloutManager(reactContext), new PolygonManager(reactContext), new PolylineManager(reactContext), new CircleManager(reactContext), new OverlayManager(reactContext));
     }
 
 
     @Override
     public NativeModule getModule(String name, ReactApplicationContext reactContext) {
+        if (OverlayManager.REACT_CLASS.equals(name)) {
+            return new OverlayManager(reactContext);
+        }
+        if (CircleManager.REACT_CLASS.equals(name)) {
+            return new CircleManager(reactContext);
+        }
+        if (PolylineManager.REACT_CLASS.equals(name)) {
+            return new PolylineManager(reactContext);
+        }
         if (PolygonManager.REACT_CLASS.equals(name)) {
             return new PolygonManager(reactContext);
         }