+++ android/src/main/java/com/rnmaps/maps/MapPolylineManager.java
@@ -74,22 +74,7 @@ public class MapPolylineManager extends ViewGroupManager<MapPolyline> {
 
   @ReactProp(name = "lineCap")
   public void setlineCap(MapPolyline view, String lineCap) {
-    Cap cap = null;
-    switch (lineCap) {
-      case "butt":
-        cap = new ButtCap();
-        break;
-      case "round":
-        cap = new RoundCap();
-        break;
-      case "square":
-        cap = new SquareCap();
-        break;
-      default:
-        cap = new RoundCap();
-        break;
-    }
-    view.setLineCap(cap);
+      view.setLineCap(lineCap);
   }
 
   @ReactProp(name = "lineDashPattern")