+++ src/decorateMapComponent.ts
@@ -23,6 +23,9 @@ import {Commands} from './MapViewNativeComponent';
 import GooglePolygon from './specs/NativeComponentGooglePolygon';
 import FabricMarker from './specs/NativeComponentMarker';
 import FabricCallout from './specs/NativeComponentCallout';
+import FabricPolyline from './specs/NativeComponentPolyline';
+import FabricCircle from './specs/NativeComponentCircle';
+import FabricOverlay from './specs/NativeComponentOverlay';
 
 export const SUPPORTED: ImplementationStatus = 'SUPPORTED';
 export const USES_DEFAULT_IMPLEMENTATION: ImplementationStatus =
@@ -88,9 +91,20 @@ export default function decorateMapComponent<Type extends Component>(
         // @ts-ignore
         return GooglePolygon;
       }
-      if (componentName === 'Callout' && Platform.OS === 'android') {
-        // @ts-ignore
-        return FabricCallout;
+      if (Platform.OS === 'android') {
+        if (componentName === 'Callout') {
+          // @ts-ignore
+          return FabricCallout;
+        } else if (componentName === 'Polyline') {
+          // @ts-ignore
+          return FabricPolyline;
+        } else if (componentName === 'Circle') {
+          // @ts-ignore
+          return FabricCircle;
+        } else if (componentName === 'Overlay') {
+          // @ts-ignore
+          return FabricOverlay;
+        }
       }
       const key = provider || 'default';
       if (components[key]) {